requirements.txt
python-dotenv
influxdb-client-python
pandas 
schedule
scikit-learn